Abstract away getting of current time in porting layer
